Renames Folders Across The Board (#23)
I don't think the CODEOWNERS bot action can accurately parse files with spaces in them, even if they are escaped. I couldn't find any functionality for it, so it keeps failing on every PR and continues to remain annoying. So, I decided to go through and make all the folders and files be one word.