Pubby Update (Cargo and More)
I weaponized the repository for a shitpost today so I thought I should actually do something productive to counter-balance it. I also wanted to test the limits of StrongDMM 2.6.0, and it came up (understandably short).

* To reflect https://github.com/tgstation/tgstation/pull/67518 , I mapped in the QM's stuff (ID trimmer, access console, red alert console).

* The bar atrium is now the lounge.

* Fixed some reinforced walls around Cargo.

* Weird glass-placement of structures moved to walls (HOPLine, Cargo).

* Bar Maintenance Renamed to Central Maintenance (clarity while mapping, better name, etc.)

* Captain's Office is now it's own area (Private Quarters) to reduce blobbosity.

* Firing Range in Brig Maintenance is now it's own area for blobbosity reduction.

* Door ownership changes in maintenance.

* Tcommsat Injector now a passive vent

* Tweaked internal security of the tcommsat to be normal walls inside, it doesn't really need to all be reinforced.

* Tcommsat Camera Areas Fixed (were in wrong area (nearstation))

* Turned the "cockfight" room in engineering maintenance to the "abandoned gaming den" for blobbosity

* Probably some other stuff (unlikely).

One day I'll muster up the willpower to map in the freezer room for Ordnance, but today is not that day. It feels good to just go through an entire map without a pre-planned scope and just fix/alter stuff.