MapDiffBot2 Experimentations
I'm going to do this PR such that I add both add it to a milestone, and so I can discuss later commits when I merge them in this one centralized place for easier reference later on. Alright, cool.

So, I think MDB2 can't find the .dme, which is to be expected. Let's make a new DME and see if we can't somehow get it to tick the tgstation.dme in the submodule.

It compiles, but it barely works in DM. I'm happy with the way it turned out since it did seem to grab everything needed. I'm stealing the .gitignore from tgstation/tgstation as well by the way.