Adds text files with the Commit Histories to Pubby and Disc
Hey there,

I believe paramount to anything, you should have some sort of honorary reference to the work that you have contributed. One failing of the map_depot is that we aren't able to perfectly import the commit history (would be a lot of work that might not even work), but I just decided to do something nice to commit all those souls who worked towards something. I think it's something that I would want done for whatever contributions I do to whatever maps we get slamdunked.

This text file that I've generated for both Pubby and Disc contains the specific /tg/ base commit, the author's name, the name of the PR, and the PR's ID. I really like this because it's a great example of all the work that have been done on this maps well before many of us arrived, and serves as a small reference point as we continue to uptain certain maps.

I just ran `git log --pretty=format:"%h%x09%an%x09%d%x09%s" --graph (map file here) > mapname_history.txt` on a branch where I had checked out only the very last version of it being present on tgstation/tgstation (not the base commit where it was removed!). I like the way it turned out.

I'm only doing it for Pubby and Disc because everything else is outdated and these are the only two that have other people who actively care about them (which makes me care about them).