Submodule File Change Helpful CI Popup Comment (#63)
* Codeowner Reviews Helpful Popup

The intrusive thoughts won today, so I thought I would take a step to ensure that we don't get absolutely rammed should no one realize we shouldn't touch that. I'm not going to be around forever, so hopefully someone who takes the reins of this after me realizes that it should be changed otherwise stuff might york. I just want to ensure everyone's on the latest submodule, and they don't try and do bullshittery.