Update Ministation to February 2023 (#75)
What it says on the tin.

Someone asked for help getting ministation to run last week and I ran it
through the update scripts. Since I already put in the work, I figured I
may as well share it here. I put in another couple hours and cleaned up
all the access varedits and pipenet issues to get rid of all the
runtimes.