Added Ministation variant (#77)
Added a variant of Ministation, developed by my friends and I for a
small private server. We're currently in the middle of a major overhaul
of the map, but figured this was a good time to post our current
version. A bit janky due to trying to fit all jobs into a tiny map, but
has been playtested continuously over the course of two years.