Adds a modernized version of "OmegaStation" (#92)
I worked on modernizing OmegaStation for another server Just4Fun and
thought I would drop the finished product here.

This is a version of OmegaStation, slightly modernized. Most
importantly, it is post-smart-pipes and post-smart-wires, while the
other version is not, making it extremely hard to work with now-a-days.

However, I have left the older version still in the repo, because this
version takes some liberties and gives Omega facilities for all jobs
(and all heads of staff), making it more "feature complete" but less
"true to form".

![StrongDMM-2024-04-13 12 19
08](https://github.com/tgstation/map_depot/assets/51863163/ceafc01e-4cb3-4e43-bfb3-8f27ac5a3e3e)

Credit is to be given to the maintainers of Citadel Station. This
version of OmegaStation was built off of their version, which is
modernized in some ways (More jobs, many paths updated) but still had no
smart wires or pipes. I had to do all those anew.