Fix Cerestation Json for Custom Use (#105)
This **just** fixes the json file so that it can be run as a custom map
without silently failing or hanging the custom map uploader.

This map probably should not be run on a live server as there are a
very, large numbers of errors. This includes pre-access-helper,
pre-list-access var definitions on doors and broken area definitions.
This is throwing 300+ runtimes got me, several of which are for
important things. Large amounts of game content may be missing or
broken.