Update Northstar (#107)
Updates Northstar through the current update path scripts.

This is not currently playable due to the deletion of Northstar's custom
shuttles in https://github.com/tgstation/tgstation/pull/87937.

Currently loading this map will result in a broken cargo department
(creating a large volume of runtimes) and latejoin players being sent to
the error room due to the lack of the arrivals shuttle. Discussion with
the other Maptainers in progress to determine how to resolve this.