updated phobos (#124)
this is the version of phobos ive been workin on and off, keeping it
mostly up to date with new station things that have been added and
fixing mistakes that crop up.
I do not remember specifically what i've done since this was over a
period of months when i would run the map for fun on the servers.

---------

Co-authored-by: Vire <66576896+Maurukas@users.noreply.github.com>