Adds monastery cameras to Kilo (#129)
Like with Pubby, Kilo also has a monastery. This adds the proper
monastery cameras & camera consoles. One outside the pipes and one to
the HoP (like Pubby), and one to the Chaplain in their office. Kilo also
has an external airlock but doesn't seem to have a destination for any
shuttle to dock on it, which on Pubby the one at dorms has a monastery
camera console next to it, but due to the lack of destination I excluded
it from Kilo.